new generated entryChoice statemachine

+++ SCT2/trunk/test-plugins/org.yakindu.sct.generator.java.test/src-gen/org/yakindu/scr/alwaysoncycle/IAlwaysOncycleStatemachine.java
@@ -4,8 +4,8 @@ import org.yakindu.scr.IStatemachine;
 public interface IAlwaysOncycleStatemachine extends IStatemachine {
 
 	public interface SCInterface {
-		public int getValue();
-		public void setValue(int value);
+		public long getValue();
+		public void setValue(long value);
 		public boolean getV2();
 		public void setV2(boolean value);
