+++ SCT2/trunk/test-plugins/org.yakindu.sct.generator.java.test/src-gen/org/yakindu/scr/exitonselftransition/IExitOnSelfTransitionStatemachine.java
@@ -6,10 +6,10 @@ public interface IExitOnSelfTransitionStatemachine extends IStatemachine {
 	public interface SCInterface {
 		public void raiseE();
 		public void raiseF();
-		public int getEntryCount();
-		public void setEntryCount(int value);
-		public int getExitCount();
-		public void setExitCount(int value);
+		public long getEntryCount();
+		public void setEntryCount(long value);
+		public long getExitCount();
+		public void setExitCount(long value);
 
 	}
