+++ SCT2/trunk/test-plugins/org.yakindu.sct.generator.java.test/src-gen/org/yakindu/scr/integerexpressions/IIntegerExpressionsStatemachine.java
@@ -5,10 +5,10 @@ public interface IIntegerExpressionsStatemachine extends IStatemachine {
 
 	public interface SCInterface {
 		public void raiseE1();
-		public int getMyInt1();
-		public void setMyInt1(int value);
-		public int getMyInt2();
-		public void setMyInt2(int value);
+		public long getMyInt1();
+		public void setMyInt1(long value);
+		public long getMyInt2();
+		public void setMyInt2(long value);
 		public boolean getLess();
 		public void setLess(boolean value);
 		public boolean getGreater();
@@ -21,30 +21,30 @@ public interface IIntegerExpressionsStatemachine extends IStatemachine {
 		public void setEqual(boolean value);
 		public boolean getNotEqual();
 		public void setNotEqual(boolean value);
-		public int getPlus();
-		public void setPlus(int value);
-		public int getMinus();
-		public void setMinus(int value);
-		public int getMultiply();
-		public void setMultiply(int value);
-		public int getDivision();
-		public void setDivision(int value);
-		public int getModulo();
-		public void setModulo(int value);
-		public int getNegat();
-		public void setNegat(int value);
+		public long getPlus();
+		public void setPlus(long value);
+		public long getMinus();
+		public void setMinus(long value);
+		public long getMultiply();
+		public void setMultiply(long value);
+		public long getDivision();
+		public void setDivision(long value);
+		public long getModulo();
+		public void setModulo(long value);
+		public long getNegat();
+		public void setNegat(long value);
 		public boolean getComplement();
 		public void setComplement(boolean value);
-		public int getMultiAssign();
-		public void setMultiAssign(int value);
-		public int getDivAssign();
-		public void setDivAssign(int value);
-		public int getPlusAssign();
-		public void setPlusAssign(int value);
-		public int getMinusAssign();
-		public void setMinusAssign(int value);
-		public int getModuloAssign();
-		public void setModuloAssign(int value);
+		public long getMultiAssign();
+		public void setMultiAssign(long value);
+		public long getDivAssign();
+		public void setDivAssign(long value);
+		public long getPlusAssign();
+		public void setPlusAssign(long value);
+		public long getMinusAssign();
+		public void setMinusAssign(long value);
+		public long getModuloAssign();
+		public void setModuloAssign(long value);
 
 	}
