+++ SCT2/trunk/test-plugins/org.yakindu.sct.generator.java.test/src-gen/org/yakindu/scr/logicalor/ILogicalOrStatemachine.java
@@ -4,8 +4,8 @@ import org.yakindu.scr.IStatemachine;
 public interface ILogicalOrStatemachine extends IStatemachine {
 
 	public interface SCInterface {
-		public int getX();
-		public void setX(int value);
+		public long getX();
+		public void setX(long value);
 		public boolean getB();
 		public void setB(boolean value);
