+++ SCT2/trunk/test-plugins/org.yakindu.sct.generator.java.test/src-gen/org/yakindu/scr/statechartlocalreactions/IStatechartLocalReactionsStatemachine.java
@@ -4,8 +4,8 @@ import org.yakindu.scr.IStatemachine;
 public interface IStatechartLocalReactionsStatemachine extends IStatemachine {
 
 	public interface SCInterface {
-		public int getMyInt();
-		public void setMyInt(int value);
+		public long getMyInt();
+		public void setMyInt(long value);
 
 	}
