+++ SCT2/trunk/test-plugins/org.yakindu.sct.generator.java.test/src-gen/org/yakindu/scr/syncjoin/ISyncJoinStatemachine.java
@@ -8,8 +8,8 @@ public interface ISyncJoinStatemachine extends IStatemachine {
 		public void raiseF();
 		public void raiseJc();
 		public void raiseJd();
-		public int getX();
-		public void setX(int value);
+		public long getX();
+		public void setX(long value);
 
 	}
