+++ SCT2/trunk/test-plugins/org.yakindu.sct.generator.java.test/src-gen/org/yakindu/scr/valuedevents/IValuedEventsStatemachine.java
@@ -4,9 +4,9 @@ import org.yakindu.scr.IStatemachine;
 public interface IValuedEventsStatemachine extends IStatemachine {
 
 	public interface SCInterface {
-		public void raiseIntegerEvent(int value);
-		public int getMyVar();
-		public void setMyVar(int value);
+		public void raiseIntegerEvent(long value);
+		public long getMyVar();
+		public void setMyVar(long value);
 
 	}
